refactor: move functions
Move functions for interacting with npm registry into client module.
